involve the community to the process of releases
In software engineering the releasing process is an important step.  Since most
contributors are comfortable enough with the submitting patches process, it is
good time to involve everybody to the releasing one.

Introduce the Changelog file to rely on it. It is the good place to put topics
not already released.

This way, releases benefit from the review of others in the same way we do for
patches.

Signed-off-by: Nicolas Sebrecht <nicolas.s-dev@laposte.net>